requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
discord.py==1.2.5
dnspython==1.16.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.18.2
multidict==4.7.2
pylint==2.4.4
pymongo==3.10.0
requests==2.22.0
six==1.13.0
tabulate==0.8.6
urllib3==1.25.7
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
